Move SerialPort class to its own Java file
