+++ java/src/jmri/jmrix/SerialPortDataListener.java
@@ -1,0 +1,14 @@
+package jmri.jmrix;
+
+/**
+ * Serial port listener
+ *
+ * @author Daniel Bergqvist (C) 2024
+ */
+public interface SerialPortDataListener {
+
+    void serialEvent(SerialPortEvent serialPortEvent);
+
+    public int getListeningEvents();
+
+}
