+++ java/src/jmri/jmrix/mrc/serialdriver/SerialDriverAdapter.java
@@ -1,5 +1,6 @@
 package jmri.jmrix.mrc.serialdriver;
 
+import jmri.jmrix.SerialPort;
 import jmri.jmrix.mrc.MrcPacketizer;
 import jmri.jmrix.mrc.MrcPortController;
 import jmri.jmrix.mrc.MrcSystemConnectionMemo;
@@ -25,7 +26,7 @@ public class SerialDriverAdapter extends MrcPortController {
     @Override
     public String openPort(String portName, String appName) {
         // get and open the primary port
-        currentSerialPort = activatePort(this.getSystemPrefix(), portName, log, 1, Parity.ODD);
+        currentSerialPort = activatePort(this.getSystemPrefix(), portName, log, 1, SerialPort.Parity.ODD);
         if (currentSerialPort == null) {
             log.error("failed to connect MRC to {}", portName);
             return Bundle.getMessage("SerialPortNotFound", portName);
