+++ java/src/jmri/jmrix/purejavacomm/CommPortIdentifier.java
@@ -19,8 +19,8 @@ public class CommPortIdentifier {
 
     public SerialPort open(String appName, int timeout) throws PortInUseException, NoSuchPortException {
         String systemPrefix = appName;
-        AbstractSerialPortController.SerialPort serialPort = AbstractSerialPortController.activatePort(
-                systemPrefix, _portName, log, 1, AbstractSerialPortController.Parity.NONE);
+        jmri.jmrix.SerialPort serialPort = AbstractSerialPortController.activatePort(
+                systemPrefix, _portName, log, 1, jmri.jmrix.SerialPort.Parity.NONE);
 
         if (serialPort != null) {
             return new SerialPort(serialPort);
