+++ java/test/jmri/ArchitectureTest.java
@@ -311,8 +311,8 @@ public class ArchitectureTest {
         .that()
 
         // all the standard serial access should be confined to here:
-        .doNotHaveFullyQualifiedName("jmri.jmrix.AbstractSerialPortController$SerialPort").and()
-        .doNotHaveFullyQualifiedName("jmri.jmrix.AbstractSerialPortController$SerialPortEvent")
+        .doNotHaveFullyQualifiedName("jmri.jmrix.SerialPort").and()
+        .doNotHaveFullyQualifiedName("jmri.jmrix.SerialPortEvent")
 
         .should().accessClassesThat().resideInAPackage("com.fazecast.jSerialComm..");
